Refactor and refinement ClickHouse compatible mode (#8169)
