+++ src/Common/ClickHouseCompatibleFlag.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace DB
+{
+bool isClickHouseCompatibleMode();
+void setClickHouseCompatibleMode(bool is_clickhouse_compatible);
+}
