--- src/Common/thread_local_is_clickhouse_compatible.cpp
@@ -1,3 +1,0 @@
-#include <Common/thread_local_is_clickhouse_compatible.h>
-
-thread_local bool thread_local_is_clickhouse_compatible = false;
