--- src/Common/thread_local_is_clickhouse_compatible.h
@@ -1,3 +1,0 @@
-#include <thread>
-
-extern thread_local bool thread_local_is_clickhouse_compatible;
