+++ src/Core/Block.h
@@ -95,9 +95,8 @@ public:
     NamesAndTypes getNamesAndTypes() const;
     Names getNames() const;
     DataTypes getDataTypes() const;
-    /// proton: starts.
-    Names getDataTypeNames(bool is_clickhouse_compatible = false) const;
-    /// proton: starts.
+    Names getDataTypeNames() const;
+    Names getClickHouseDataTypeNames() const; /// Added by proton to get ClickHouse compatible data types
     std::unordered_map<String, size_t> getNamesToIndexesMap() const;
 
     /// Returns number of rows from first column in block, not equal to nullptr. If no columns, returns 0.
