+++ src/Core/Settings.h
@@ -634,7 +634,7 @@ static constexpr UInt64 operator""_GiB(unsigned long long value)
     \
     /** proton: starts */                       \
     M(String, shards, "", "If not empty, only the specified shard IDs (or partition IDs if the target stream is a Kafka external stream) will be selected to be read data from. IDs are separated by comma. Example: shards='0,2'", 0) \
-    M(Bool, is_clickhouse_compatible, false, "When this parameter is set, timeplusd works in compatible mode with ClickHouse (especially compatible with data types and function names).", 0)  \
+    M(Bool, is_clickhouse_compatible, false, "When this setting is set to true, timeplus can run ClickHouse SQL", 0)  \
     \
     /** proton: ends */                       \
