+++ src/Parsers/ExpressionElementParsers.cpp
@@ -43,19 +43,12 @@
 #include <Interpreters/StorageID.h>
 
 /// proton: starts.
-#include <Parsers/Streaming/ParserIntervalAliasExpression.h>
 #include <Parsers/Streaming/ParserSessionRangeComparisonExpressionIfPossible.h>
+#include <Common/ClickHouseCompatibleFlag.h>
 /// proton: ends.
-
-/// proton: starts
-#include <Common/thread_local_is_clickhouse_compatible.h>
-/// proton: ends
 
 namespace DB
 {
-/// proton: starts
-class Context;
-/// proton: ends
 
 namespace ErrorCodes
 {
@@ -756,1260 +749,6 @@ namespace
     }
 }
 
-/// proton: starts
-/// When timeplus registers a new function, the new function needs to be added to one of the following two maps.
-static std::unordered_map<std::string, std::string> function_map
-    = {{"cast", "cast"},
-       {"encodeURLComponent", "encode_url_component"},
-       {"decodeURLComponent", "decode_url_component"},
-       {"URLHierarchy", "url_hierarchy"},
-       {"URLPathHierarchy", "url_path_hierarchy"},
-       {"extractURLParameterNames", "extract_url_parameter_names"},
-       {"extractURLParameters", "extract_url_parameters"},
-       {"generateUUIDv4", "generate_uuidv4"},
-       {"toYYYYMMDDhhmmss", "to_YYYYMMDDhhmmss"},
-       {"toYYYYMM", "to_YYYYMM"},
-       {"to_YYYYMM", "to_YYYYMM"},
-       {"BLAKE3", "blake3"},
-       {"connectionId", "connection_id"},
-       {"countMatches", "count_matches"},
-       {"countMatchesCaseInsensitive", "count_matches_case_insensitive"},
-       {"dateName", "date_name"},
-       {"displayName", "display_name"},
-       {"formatReadableDecimalSize", "format_readable_decimal_size"},
-       {"FQDN", "fqdn"},
-       {"hasToken", "has_token"},
-       {"leftUTF8", "left_utf8"},
-       {"position", "position"},
-       {"rightUTF8", "right_utf8"},
-       {"roundBankers", "round_bankers"},
-       {"starts", "starts"},
-       {"structureToCapnProtoSchema", "structure_to_capn_proto_schema"},
-       {"structureToProtobufSchema", "structure_to_protobuf_schema"},
-       {"xxh3", "xxh3"}};
-
-static std::unordered_map<std::string, std::string> case_insensitive_function_map
-    /// case_insensitive_functions
-    = {{"bithammingdistance", "bit_hamming_distance"},
-       {"polygonsdistancespherical", "polygons_distance_spherical"},
-       {"polygonsdistancecartesian", "polygons_distance_cartesian"},
-       {"fuzzbits", "fuzz_bits"},
-       {"initializeaggregation", "initialize_aggregation"},
-       {"extractgroups", "extract_groups"},
-       {"tostartofsecond", "to_start_of_second"},
-       {"randomintype", "random_in_type"},
-       {"jsonvalues", "json_values"},
-       {"jsonexists", "json_exists"},
-       {"rand", "rand"},
-       {"tointervalmonth", "to_interval_month"},
-       {"tointervalminute", "to_interval_minute"},
-       {"tointervalsecond", "to_interval_second"},
-       {"tointervalmillisecond", "to_interval_millisecond"},
-       {"tointervalmicrosecond", "to_interval_microsecond"},
-       {"parsedatetime64besteffortusornull", "parse_datetime64_best_effort_us_or_null"},
-       {"parsedatetime64besteffortorzero", "parse_datetime64_best_effort_or_zero"},
-       {"parsedatetime64besteffort", "parse_datetime64_best_effort"},
-       {"parsedatetime32besteffortornull", "parse_datetime32_best_effort_or_null"},
-       {"parsedatetimebesteffortusorzero", "parse_datetime_best_effort_us_or_zero"},
-       {"parsedatetimebesteffortus", "parse_datetime_best_effort_us"},
-       {"toipv6ornull", "to_ipv6_or_null"},
-       {"toipv4ornull", "to_ipv4_or_null"},
-       {"touuidornull", "to_uuid_or_null"},
-       {"todecimal32ornull", "to_decimal32_or_null"},
-       {"todate32ornull", "to_date32_or_null"},
-       {"tofloat64ornull", "to_float64_or_null"},
-       {"tofloat32ornull", "to_float32_or_null"},
-       {"toint256ornull", "to_int256_or_null"},
-       {"toint128ornull", "to_int128_or_null"},
-       {"toint64ornull", "to_int64_or_null"},
-       {"toint16ornull", "to_int16_or_null"},
-       {"touint256ornull", "to_uint256_or_null"},
-       {"touint32ornull", "to_uint32_or_null"},
-       {"touint16ornull", "to_uint16_or_null"},
-       {"touint8ornull", "to_uint8_or_null"},
-       {"toipv6orzero", "to_ipv6_or_zero"},
-       {"toipv4orzero", "to_ipv4_or_zero"},
-       {"touuidorzero", "to_uuid_or_zero"},
-       {"todecimal128orzero", "to_decimal128_or_zero"},
-       {"todecimal64orzero", "to_decimal64_or_zero"},
-       {"todecimal32orzero", "to_decimal32_or_zero"},
-       {"todatetimeorzero", "to_datetime_or_zero"},
-       {"todate32orzero", "to_date32_or_zero"},
-       {"tofloat64orzero", "to_float64_or_zero"},
-       {"toint64orzero", "to_int64_or_zero"},
-       {"toint8orzero", "to_int8_or_zero"},
-       {"parsedatetimebesteffortorzero", "parse_datetime_best_effort_or_zero"},
-       {"touint128orzero", "to_uint128_or_zero"},
-       {"touint64orzero", "to_uint64_or_zero"},
-       {"touint32orzero", "to_uint32_or_zero"},
-       {"touint8orzero", "to_uint8_or_zero"},
-       {"toipv4", "to_ipv4"},
-       {"todatetime64", "to_datetime64"},
-       {"todate", "to_date"},
-       {"todate16", "to_date16"},
-       {"todecimal128", "to_decimal128"},
-       {"todecimal64", "to_decimal64"},
-       {"todecimal32", "to_decimal32"},
-       {"toint256", "to_int256"},
-       {"toint128", "to_int128"},
-       {"toint64", "to_int64"},
-       {"toint32", "to_int32"},
-       {"touint256", "to_uint256"},
-       {"touint128", "to_uint128"},
-       {"touint64", "to_uint64"},
-       {"touint32", "to_uint32"},
-       {"touint16", "to_uint16"},
-       {"touint8", "to_uint8"},
-       {"blocksize", "block_size"},
-       {"isnotnull", "is_not_null"},
-       {"rounddown", "round_down"},
-       {"trunc", "trunc"},
-       {"floor", "floor"},
-       {"round", "round"},
-       {"weakhash32", "weak_hash32"},
-       {"wyhash64", "wy_hash64"},
-       {"xxhash64", "xx_hash64"},
-       {"xxhash32", "xx_hash32"},
-       {"gccmurmurhash", "gcc_murmur_hash"},
-       {"murmurhash3128", "murmur_hash3_128"},
-       {"murmurhash364", "murmur_hash3_64"},
-       {"todecimal256orzero", "to_decimal256_or_zero"},
-       {"murmurhash332", "murmur_hash3_32"},
-       {"hivehash", "hive_hash"},
-       {"javahashutf16le", "java_hash_utf16_le"},
-       {"javahash", "java_hash"},
-       {"urlhash", "url_hash"},
-       {"inthash64", "int_hash64"},
-       {"inthash32", "int_hash32"},
-       {"metrohash64", "metro_hash64"},
-       {"farmfingerprint64", "farm_fingerprint64"},
-       {"siphash128", "sip_hash128"},
-       {"sha512", "sha512"},
-       {"sha384", "sha384"},
-       {"uuidstringtonum", "uuid_string_to_num"},
-       {"sha224", "sha224"},
-       {"sha1", "sha1"},
-       {"todecimal", "to_decimal"},
-       {"md5", "md5"},
-       {"halfmd5", "half_md5"},
-       {"lowcardinalityindices", "low_cardinality_indices"},
-       {"logtrace", "log_trace"},
-       {"edition", "edition"},
-       {"stringtoh3", "string_to_h3"},
-       {"toint16", "to_int16"},
-       {"jsonextractkeysandvaluesraw", "json_extract_keys_and_values_raw"},
-       {"jsonextractarrayraw", "json_extract_array_raw"},
-       {"jsonextractraw", "json_extract_raw"},
-       {"jsonextractkeysandvalues", "json_extract_keys_and_values"},
-       {"jsonextract", "json_extract"},
-       {"jsonextractstring", "json_extract_string"},
-       {"jsonextractbool", "json_extract_bool"},
-       {"jsonextractint", "json_extract_int"},
-       {"jsontype", "json_type"},
-       {"jsonlength", "json_length"},
-       {"isvalidjson", "is_valid_json"},
-       {"char", "char"},
-       {"cbrt", "cbrt"},
-       {"s2togeo", "s2_to_geo"},
-       {"sleepeachrow", "sleep_each_row"},
-       {"regiontoname", "region_to_name"},
-       {"regionin", "region_in"},
-       {"regiontocountry", "region_to_country"},
-       {"regiontodistrict", "region_to_district"},
-       {"adddays", "add_days"},
-       {"countsubstringscaseinsensitiveutf8", "count_substrings_case_insensitive_utf8"},
-       {"makedatetime64", "make_datetime64"},
-       {"date", "date"},
-       {"makedate", "make_date"},
-       {"subtractnanoseconds", "subtract_nanoseconds"},
-       {"ipv6stringtonumornull", "ipv6_string_to_num_or_null"},
-       {"ipv6stringtonumordefault", "ipv6_string_to_num_or_default"},
-       {"ipv6stringtonum", "ipv6_string_to_num"},
-       {"ipv6numtostring", "ipv6_num_to_string"},
-       {"ipv4stringtonumordefault", "ipv4_string_to_num_or_default"},
-       {"ipv4stringtonum", "ipv4_string_to_num"},
-       {"ipv4numtostringclassc", "ipv4_num_to_string_class_c"},
-       {"isipv4string", "is_ipv4_string"},
-       {"ipv4cidrtorange", "ipv4_cidr_to_range"},
-       {"ipv6cidrtorange", "ipv6_cidr_to_range"},
-       {"macstringtooui", "mac_string_to_oui"},
-       {"macnumtostring", "mac_num_to_string"},
-       {"ipv4toipv6", "ipv4_to_ipv6"},
-       {"cutipv6", "cut_ipv6"},
-       {"snowflaketodatetime64", "snowflake_to_datetime64"},
-       {"subtractmilliseconds", "subtract_milliseconds"},
-       {"getserverport", "get_server_port"},
-       {"regiontocity", "region_to_city"},
-       {"multisearchany", "multi_search_any"},
-       {"bitmaphasany", "bitmap_has_any"},
-       {"bitmapandnot", "bitmap_andnot"},
-       {"parsedatetime64besteffortus", "parse_datetime64_best_effort_us"},
-       {"bitmapand", "bitmap_and"},
-       {"detectcharset", "detect_charset"},
-       {"bitmapxorcardinality", "bitmap_xor_cardinality"},
-       {"h3getpentagonindexes", "h3_get_pentagon_indexes"},
-       {"arraylastindex", "array_last_index"},
-       {"bitmaporcardinality", "bitmap_or_cardinality"},
-       {"subbitmap", "sub_bitmap"},
-       {"runningconcurrency", "running_concurrency"},
-       {"bitmapsubsetlimit", "bitmap_subset_limit"},
-       {"bitmapsubsetinrange", "bitmap_subset_in_range"},
-       {"bitmapbuild", "bitmap_build"},
-       {"extracttextfromhtml", "extract_text_from_html"},
-       {"monthname", "month_name"},
-       {"splitbyregexp", "split_by_regexp"},
-       {"isnan", "is_nan"},
-       {"geohashencode", "geohash_encode"},
-       {"torelativequarternum", "to_relative_quarter_num"},
-       {"or", "or"},
-       {"dictgetchildren", "dict_get_children"},
-       {"dictisin", "dict_is_in"},
-       {"todecimal128ornull", "to_decimal128_or_null"},
-       {"dictgetstringordefault", "dict_get_string_or_default"},
-       {"dictgetipv6ordefault", "dict_get_ipv6_or_default"},
-       {"dictgetdatetimeordefault", "dict_get_datetime_or_default"},
-       {"datetimetosnowflake", "datetime_to_snowflake"},
-       {"todateordefault", "to_date_or_default"},
-       {"dictgetdateordefault", "dict_get_date_or_default"},
-       {"dictgetfloat64ordefault", "dict_get_float64_or_default"},
-       {"h3getresolution", "h3_get_resolution"},
-       {"exp", "exp"},
-       {"dictgetfloat32ordefault", "dict_get_float32_or_default"},
-       {"addseconds", "add_seconds"},
-       {"dictgetint64ordefault", "dict_get_int64_or_default"},
-       {"toint", "to_int"},
-       {"dictgetint16ordefault", "dict_get_int16_or_default"},
-       {"dictgetuint64ordefault", "dict_get_uint64_or_default"},
-       {"emptyarraystring", "empty_array_string"},
-       {"dictgetuint32ordefault", "dict_get_uint32_or_default"},
-       {"h3cellaream2", "h3_cell_area_m2"},
-       {"date_diff", "date_diff"},
-       {"dictgetstring", "dict_get_string"},
-       {"dictgetuuid", "dict_get_uuid"},
-       {"dictgetfloat64", "dict_get_float64"},
-       {"tuplemultiplybynumber", "tuple_multiply_by_number"},
-       {"tomonth", "to_month"},
-       {"dictgetint16", "dict_get_int16"},
-       {"dictgetint8", "dict_get_int8"},
-       {"dictgetuint32", "dict_get_uint32"},
-       {"multisearchanycaseinsensitiveutf8", "multi_search_any_case_insensitive_utf8"},
-       {"cosh", "cosh"},
-       {"dictgetuint16", "dict_get_uint16"},
-       {"sleep", "sleep"},
-       {"dictget", "dict_get"},
-       {"replicate", "replicate"},
-       {"tomodifiedjulianday", "to_modified_julian_day"},
-       {"h3getunidirectionaledge", "h3_get_unidirectional_edge"},
-       {"regiontocontinent", "region_to_continent"},
-       {"hascolumnintable", "has_column_in_table"},
-       {"detectlanguageunknown", "detect_language_unknown"},
-       {"arrayslice", "array_slice"},
-       {"todecimal128ordefault", "to_decimal128_or_default"},
-       {"partitionid", "partition_id"},
-       {"formatdatetimeinjodasyntax", "format_datetime_in_joda_syntax"},
-       {"dictgetornull", "dict_get_or_null"},
-       {"tostartofquarter", "to_start_of_quarter"},
-       {"multisearchfirstindex", "multi_search_first_index"},
-       {"bytesize", "byte_size"},
-       {"geohashdecode", "geohash_decode"},
-       {"polygonsequalscartesian", "polygons_equals_cartesian"},
-       {"jsonkey", "json_key"},
-       {"bitmapmax", "bitmap_max"},
-       {"tid", "tid"},
-       {"bitcount", "bit_count"},
-       {"tointervalyear", "to_interval_year"},
-       {"not", "not"},
-       {"formatrow", "format_row"},
-       {"arrayreduceinranges", "array_reduce_in_ranges"},
-       {"addmicroseconds", "add_microseconds"},
-       {"geotoh3", "geo_to_h3"},
-       {"bitmapmin", "bitmap_min"},
-       {"normalizequery", "normalize_query"},
-       {"if", "if"},
-       {"defaultprofiles", "default_profiles"},
-       {"countsubstrings", "count_substrings"},
-       {"ngramdistancecaseinsensitiveutf8", "ngram_distance_case_insensitive_utf8"},
-       {"acosh", "acosh"},
-       {"accuratecastornull", "accurate_cast_or_null"},
-       {"h3indexesareneighbors", "h3_indexes_are_neighbors"},
-       {"accuratecast", "accurate_cast"},
-       {"cast", "cast"},
-       {"notlike", "not_like"},
-       {"aesdecryptmysql", "aes_decrypt_mysql"},
-       {"toint8", "to_int8"},
-       {"dictgetfloat32", "dict_get_float32"},
-       {"arraycompact", "array_compact"},
-       {"dictgetint32", "dict_get_int32"},
-       {"bitslice", "bit_slice"},
-       {"bitrotateright", "bit_rotate_right"},
-       {"toweek", "to_week"},
-       {"hashid", "hashid"},
-       {"h3edgelengthm", "h3_edge_length_m"},
-       {"bitshiftright", "bit_shift_right"},
-       {"tan", "tan"},
-       {"substring", "substring"},
-       {"tostartofday", "to_start_of_day"},
-       {"addhours", "add_hours"},
-       {"ipv4numtostring", "ipv4_num_to_string"},
-       {"randcanonical", "rand_canonical"},
-       {"h3togeoboundary", "h3_to_geo_boundary"},
-       {"blocknumber", "block_number"},
-       {"l2normalize", "l2_normalize"},
-       {"reverse", "reverse"},
-       {"multisearchfirstindexutf8", "multi_search_first_index_utf8"},
-       {"torelativesecondnum", "to_relative_second_num"},
-       {"greatcircleangle", "great_circle_angle"},
-       {"subtractyears", "subtract_years"},
-       {"isvalidutf8", "is_valid_utf8"},
-       {"arrayfold", "array_fold"},
-       {"casewithexpression", "case_with_expression"},
-       {"addyears", "add_years"},
-       {"tointervalquarter", "to_interval_quarter"},
-       {"tonullable", "to_nullable"},
-       {"intexp2", "int_exp2"},
-       {"wordshingleminhashargutf8", "word_shingle_min_hash_arg_utf8"},
-       {"fromunixtimestamp", "from_unix_timestamp"},
-       {"dictgetint32ordefault", "dict_get_int32_or_default"},
-       {"wordshingleminhashcaseinsensitive", "word_shingle_min_hash_case_insensitive"},
-       {"rownumberinblock", "row_number_in_block"},
-       {"splitbywhitespace", "split_by_whitespace"},
-       {"wordshingleminhasharg", "word_shingle_min_hash_arg"},
-       {"s2capcontains", "s2_cap_contains"},
-       {"ngramminhashargcaseinsensitiveutf8", "ngram_min_hash_arg_case_insensitive_utf8"},
-       {"extractallgroupsvertical", "extract_all_groups_vertical"},
-       {"globalnotnullin", "global_not_null_in"},
-       {"ngramminhashargutf8", "ngram_min_hash_arg_utf8"},
-       {"ngramminhashargcaseinsensitive", "ngram_min_hash_arg_case_insensitive"},
-       {"ngramminhashutf8", "ngram_min_hash_utf8"},
-       {"ngramminhash", "ngram_min_hash"},
-       {"wordshinglesimhashutf8", "word_shingle_sim_hash_utf8"},
-       {"jsonvalue", "json_value"},
-       {"wordshinglesimhashcaseinsensitive", "word_shingle_sim_hash_case_insensitive"},
-       {"getmacro", "get_macro"},
-       {"wordshinglesimhash", "word_shingle_sim_hash"},
-       {"ngramsimhashcaseinsensitiveutf8", "ngram_sim_hash_case_insensitive_utf8"},
-       {"ngramsimhashutf8", "ngram_sim_hash_utf8"},
-       {"wordshingleminhashargcaseinsensitiveutf8", "word_shingle_min_hash_arg_case_insensitive_utf8"},
-       {"h3pointdistrads", "h3_point_dist_rads"},
-       {"runningdifferencestartingwithfirstvalue", "running_difference_starting_with_first_value"},
-       {"ngramsimhashcaseinsensitive", "ngram_sim_hash_case_insensitive"},
-       {"base58decode", "base58_decode"},
-       {"moduloorzero", "modulo_or_zero"},
-       {"normalizeutf8nfd", "normalize_utf8_nfd"},
-       {"currentdatabase", "current_database"},
-       {"streamingnow64", "__streaming_now64"},
-       {"identity", "identity"},
-       {"ascii", "ascii"},
-       {"currentprofiles", "current_profiles"},
-       {"lowerutf8", "lower_utf8"},
-       {"polygonperimetercartesian", "polygon_perimeter_cartesian"},
-       {"positivemodulo", "positive_modulo"},
-       {"bitpositionstoarray", "bit_positions_to_array"},
-       {"dictgetipv4ordefault", "dict_get_ipv4_or_default"},
-       {"jsonhas", "json_has"},
-       {"currentuser", "current_user"},
-       {"tofloat64", "to_float64"},
-       {"age", "age"},
-       {"regionhierarchy", "region_hierarchy"},
-       {"reinterpret", "reinterpret"},
-       {"todayofweek", "to_day_of_week"},
-       {"unbin", "unbin"},
-       {"h3edgeangle", "h3_edge_angle"},
-       {"modulo", "modulo"},
-       {"toipv6ordefault", "to_ipv6_or_default"},
-       {"cos", "cos"},
-       {"unhex", "unhex"},
-       {"hex", "hex"},
-       {"gcd", "gcd"},
-       {"jumpconsistenthash", "jump_consistent_hash"},
-       {"defaultvalueofargumenttype", "default_value_of_argument_type"},
-       {"subtractminutes", "subtract_minutes"},
-       {"frommodifiedjulianday", "from_modified_julian_day"},
-       {"concat", "concat"},
-       {"demangle", "demangle"},
-       {"log", "log"},
-       {"alphatokens", "alpha_tokens"},
-       {"l1norm", "l1_norm"},
-       {"dividedecimal", "divide_decimal"},
-       {"h3tochildren", "h3_to_children"},
-       {"fullhostname", "full_host_name"},
-       {"getsetting", "get_setting"},
-       {"bitboolmaskand", "__bit_bool_mask_and"},
-       {"tupleplus", "tuple_plus"},
-       {"minus", "minus"},
-       {"sqrt", "sqrt"},
-       {"casewithoutexpr", "case_without_expr"},
-       {"bitrotateleft", "bit_rotate_left"},
-       {"getsizeofenumtype", "get_size_of_enum_type"},
-       {"endswith", "ends_with"},
-       {"wordshinglesimhashcaseinsensitiveutf8", "word_shingle_sim_hash_case_insensitive_utf8"},
-       {"streamingrownumber", "__streaming_row_number"},
-       {"formatreadablesize", "format_readable_size"},
-       {"countdigits", "count_digits"},
-       {"subtractquarters", "subtract_quarters"},
-       {"polygonareacartesian", "polygon_area_cartesian"},
-       {"trybase58decode", "try_base58_decode"},
-       {"exp10", "exp10"},
-       {"bitswaplasttwo", "__bit_swap_last_two"},
-       {"lowcardinalitykeys", "low_cardinality_keys"},
-       {"subtractinterval", "subtract_interval"},
-       {"finalizeaggregation", "finalize_aggregation"},
-       {"trybase64decode", "try_base64_decode"},
-       {"tokens", "tokens"},
-       {"ngrams", "ngrams"},
-       {"geohashesinbox", "geohashes_in_box"},
-       {"runningdifference", "running_difference"},
-       {"arrayenumeratedenseranked", "array_enumerate_dense_ranked"},
-       {"flattentuple", "flatten_tuple"},
-       {"bitmapor", "bitmap_or"},
-       {"concatassumeinjective", "concat_assume_injective"},
-       {"lcm", "lcm"},
-       {"mapextractkeylike", "map_extract_key_like"},
-       {"formatreadablequantity", "format_readable_quantity"},
-       {"formatrownonewline", "format_row_no_newline"},
-       {"casewithoutexpression", "case_without_expression"},
-       {"multiif", "multi_if"},
-       {"fromunixtimestamp64milli", "from_unix_timestamp64_milli"},
-       {"hastokencaseinsensitive", "has_token_case_insensitive"},
-       {"geodistance", "geo_distance"},
-       {"h3getres0indexes", "h3_get_res0_indexes"},
-       {"toyear", "to_year"},
-       {"decodexmlcomponent", "decode_xml_component"},
-       {"now64", "now64"},
-       {"multimatchany", "multi_match_any"},
-       {"h3exactedgelengthkm", "h3_exact_edge_length_km"},
-       {"arraysum", "array_sum"},
-       {"toisoyear", "to_iso_year"},
-       {"enabledroles", "enabled_roles"},
-       {"parsedatetimebesteffortusornull", "parse_datetime_best_effort_us_or_null"},
-       {"tostartoftenminutes", "to_start_of_ten_minutes"},
-       {"lower", "lower"},
-       {"h3exactedgelengthm", "h3_exact_edge_length_m"},
-       {"bitmapandcardinality", "bitmap_and_cardinality"},
-       {"dictgetdatetime", "dict_get_datetime"},
-       {"totime", "__to_time"},
-       {"h3getdestinationindexfromunidirectionaledge", "h3_get_destination_index_from_unidirectional_edge"},
-       {"roundduration", "round_duration"},
-       {"h3getoriginindexfromunidirectionaledge", "h3_get_origin_index_from_unidirectional_edge"},
-       {"crc32", "crc32"},
-       {"trimright", "trim_right"},
-       {"trimleft", "trim_left"},
-       {"dumpcolumnstructure", "dump_column_structure"},
-       {"touuid", "to_uuid"},
-       {"extractallgroupshorizontal", "extract_all_groups_horizontal"},
-       {"uniqthetaunion", "uniq_theta_union"},
-       {"s2cellsintersect", "s2_cells_intersect"},
-       {"coalesce", "coalesce"},
-       {"todateorzero", "to_date_or_zero"},
-       {"initialqueryid", "initial_query_id"},
-       {"toyearweek", "to_year_week"},
-       {"detectlanguagemixed", "detect_language_mixed"},
-       {"h3line", "h3_line"},
-       {"atanh", "atanh"},
-       {"wordshingleminhashutf8", "word_shingle_min_hash_utf8"},
-       {"toint32ornull", "to_int32_or_null"},
-       {"simplejsonextractuint", "simple_json_extract_uint"},
-       {"h3hexring", "h3_hex_ring"},
-       {"todecimal256", "to_decimal256"},
-       {"ngramsearch", "ngram_search"},
-       {"arraydifference", "array_difference"},
-       {"arraypushback", "array_push_back"},
-       {"h3unidirectionaledgeisvalid", "h3_unidirectional_edge_is_valid"},
-       {"arraystringconcat", "array_string_concat"},
-       {"repeat", "repeat"},
-       {"stem", "stem"},
-       {"totypename", "to_type_name"},
-       {"todatetime64ornull", "to_datetime64_or_null"},
-       {"reversednsquery", "reverse_dns_query"},
-       {"splitbynonalpha", "split_by_non_alpha"},
-       {"ignore", "ignore"},
-       {"tostartofnanosecond", "to_start_of_nanosecond"},
-       {"bittest", "bit_test"},
-       {"tointervalhour", "to_interval_hour"},
-       {"h3toparent", "h3_to_parent"},
-       {"buildid", "build_id"},
-       {"multisearchfirstposition", "multi_search_first_position"},
-       {"dictgetuint64", "dict_get_uint64"},
-       {"h3hexareakm2", "h3_hex_area_km2"},
-       {"h3kring", "h3k_ring"},
-       {"toint8ornull", "to_int8_or_null"},
-       {"hypot", "hypot"},
-       {"kostikconsistenthash", "kostik_consistent_hash"},
-       {"globalnotnullinignoreset", "global_not_null_in_ignore_set"},
-       {"detectlanguage", "detect_language"},
-       {"tupleelement", "tuple_element"},
-       {"notnullinignoreset", "not_null_in_ignore_set"},
-       {"h3exactedgelengthrads", "h3_exact_edge_length_rads"},
-       {"globalnotinignoreset", "global_not_in_ignore_set"},
-       {"globalinignoreset", "global_in_ignore_set"},
-       {"subtractmonths", "subtract_months"},
-       {"inignoreset", "in_ignore_set"},
-       {"bitshiftleft", "bit_shift_left"},
-       {"notnullin", "not_null_in"},
-       {"globalnullin", "global_null_in"},
-       {"globalnotin", "global_not_in"},
-       {"bitmaptransform", "bitmap_transform"},
-       {"parsetimedelta", "parse_time_delta"},
-       {"extractkeyvaluepairs", "extract_key_value_pairs"},
-       {"in", "in"},
-       {"h3getbasecell", "h3_get_base_cell"},
-       {"errorcodetoname", "error_code_to_name"},
-       {"randomfixedstring", "random_fixed_string"},
-       {"tofloat32orzero", "to_float32_or_zero"},
-       {"acos", "acos"},
-       {"isconstant", "is_constant"},
-       {"countsubstringscaseinsensitive", "count_substrings_case_insensitive"},
-       {"emptyarraytosingle", "empty_array_to_single"},
-       {"isipv6string", "is_ipv6_string"},
-       {"addnanoseconds", "add_nanoseconds"},
-       {"monotonic", "monotonic"},
-       {"randexponential", "rand_exponential"},
-       {"extractkeyvaluepairswithescaping", "extract_key_value_pairs_with_escaping"},
-       {"polygonsunionspherical", "polygons_union_spherical"},
-       {"polygonsunioncartesian", "polygons_union_cartesian"},
-       {"isnullable", "is_nullable"},
-       {"iszeroornull", "is_zero_or_null"},
-       {"macstringtonum", "mac_string_to_num"},
-       {"leftpadutf8", "left_pad_utf8"},
-       {"lagbehind", "lag_behind"},
-       {"multifuzzymatchanyindex", "multi_fuzzy_match_any_index"},
-       {"lemmatize", "lemmatize"},
-       {"isdecimaloverflow", "is_decimal_overflow"},
-       {"multisearchallpositionscaseinsensitiveutf8", "multi_search_all_positions_case_insensitive_utf8"},
-       {"lengthutf8", "length_utf8"},
-       {"h3hexaream2", "h3_hex_area_m2"},
-       {"tobool", "to_bool"},
-       {"lessorequals", "less_or_equals"},
-       {"multiplydecimal", "multiply_decimal"},
-       {"wkt", "wkt"},
-       {"filesystemunreserved", "filesystem_unreserved"},
-       {"randnormal", "rand_normal"},
-       {"abs", "abs"},
-       {"and", "and"},
-       {"arraymap", "array_map"},
-       {"log1p", "log1p"},
-       {"fromunixtimestamp64nano", "from_unix_timestamp64_nano"},
-       {"todecimal32ordefault", "to_decimal32_or_default"},
-       {"todatetime64ordefault", "to_datetime64_or_default"},
-       {"greatcircledistance", "great_circle_distance"},
-       {"reinterpretasfixedstring", "reinterpret_as_fixed_string"},
-       {"dotproduct", "dot_product"},
-       {"touint256orzero", "to_uint256_or_zero"},
-       {"jsonextractfloat", "json_extract_float"},
-       {"tofloat64ordefault", "to_float64_or_default"},
-       {"toint256ordefault", "to_int256_or_default"},
-       {"tuplenegate", "tuple_negate"},
-       {"arrayflatten", "array_flatten"},
-       {"materialize", "materialize"},
-       {"toint8ordefault", "to_int8_or_default"},
-       {"tostartofminute", "to_start_of_minute"},
-       {"extract", "extract"},
-       {"touint64ordefault", "to_uint64_or_default"},
-       {"arrayenumerateuniq", "array_enumerate_uniq"},
-       {"svg", "svg"},
-       {"bitmaphasall", "bitmap_has_all"},
-       {"h3ispentagon", "h3_is_pentagon"},
-       {"mapupdate", "map_update"},
-       {"proportionsztest", "proportions_ztest"},
-       {"base58encode", "base58_encode"},
-       {"h3isvalid", "h3_is_valid"},
-       {"bitmasktolist", "bitmask_to_list"},
-       {"mapvalues", "map_values"},
-       {"positioncaseinsensitive", "position_case_insensitive"},
-       {"mapcontains", "map_contains"},
-       {"meilimatch", "meili_match"},
-       {"snowflaketodatetime", "snowflake_to_datetime"},
-       {"multisearchfirstpositioncaseinsensitive", "multi_search_first_position_case_insensitive"},
-       {"minsamplesizeconversion", "min_sample_size_conversion"},
-       {"divide", "divide"},
-       {"modulolegacy", "modulo_legacy"},
-       {"tojsonstring", "to_json_string"},
-       {"isinfinite", "is_infinite"},
-       {"mortondecode", "morton_decode"},
-       {"enabledprofiles", "enabled_profiles"},
-       {"grok", "grok"},
-       {"mortonencode", "morton_encode"},
-       {"randomprintableascii", "random_printable_ascii"},
-       {"todatetimeordefault", "to_datetime_or_default"},
-       {"timeslot", "time_slot"},
-       {"formatreadabletimedelta", "format_readable_time_delta"},
-       {"s2rectadd", "s2_rect_add"},
-       {"runningaccumulate", "running_accumulate"},
-       {"hasthreadfuzzer", "has_thread_fuzzer"},
-       {"appendtrailingcharifabsent", "append_trailing_char_if_absent"},
-       {"bitboolmaskor", "__bit_bool_mask_or"},
-       {"multifuzzymatchany", "multi_fuzzy_match_any"},
-       {"torelativeweeknum", "to_relative_week_num"},
-       {"dictgethierarchy", "dict_get_hierarchy"},
-       {"aesencryptmysql", "aes_encrypt_mysql"},
-       {"intdivorzero", "int_div_or_zero"},
-       {"rightpadutf8", "right_pad_utf8"},
-       {"reverseutf8", "reverse_utf8"},
-       {"leftpad", "left_pad"},
-       {"translateutf8", "translate_utf8"},
-       {"multimatchanyindex", "multi_match_any_index"},
-       {"todecimal64ornull", "to_decimal64_or_null"},
-       {"arrayfirstornull", "array_first_or_null"},
-       {"addmonths", "add_months"},
-       {"timeslots", "time_slots"},
-       {"toipv4ordefault", "to_ipv4_or_default"},
-       {"joinget", "join_get"},
-       {"base64decode", "base64_decode"},
-       {"atan2", "atan2"},
-       {"toint16ordefault", "to_int16_or_default"},
-       {"lpdistance", "lp_distance"},
-       {"normalizeutf8nfkd", "normalize_utf8_nfkd"},
-       {"multisearchallpositions", "multi_search_all_positions"},
-       {"dictgetuint8", "dict_get_uint8"},
-       {"accuratecastordefault", "accurate_cast_or_default"},
-       {"file", "file"},
-       {"normalizeutf8nfkc", "normalize_utf8_nfkc"},
-       {"arraypushfront", "array_push_front"},
-       {"normalizeutf8nfc", "normalize_utf8_nfc"},
-       {"mapkeys", "map_keys"},
-       {"todatetimeornull", "to_datetime_or_null"},
-       {"encrypt", "encrypt"},
-       {"arrayauc", "array_auc"},
-       {"tostartofmonth", "to_start_of_month"},
-       {"siphash64", "sip_hash64"},
-       {"arrayelement", "array_element"},
-       {"bitor", "bit_or"},
-       {"formatdatetime", "format_datetime"},
-       {"simplejsonextractraw", "simple_json_extract_raw"},
-       {"subtractseconds", "subtract_seconds"},
-       {"addresstosymbol", "address_to_symbol"},
-       {"emitversion", "emit_version"},
-       {"tostring", "to_string"},
-       {"makedate32", "make_date32"},
-       {"simplejsonextractbool", "simple_json_extract_bool"},
-       {"bittestall", "bit_test_all"},
-       {"reinterpretasdate", "reinterpret_as_date"},
-       {"getoskernelversion", "get_os_kernel_version"},
-       {"arraycount", "array_count"},
-       {"evalmlmethod", "eval_ml_method"},
-       {"addminutes", "add_minutes"},
-       {"sin", "sin"},
-       {"positioncaseinsensitiveutf8", "position_case_insensitive_utf8"},
-       {"multisearchfirstindexcaseinsensitiveutf8", "multi_search_first_index_case_insensitive_utf8"},
-       {"mapcast", "map_cast"},
-       {"todatetime64orzero", "to_datetime64_or_zero"},
-       {"timediff", "time_diff"},
-       {"has", "has"},
-       {"base64encode", "base64_encode"},
-       {"s2rectcontains", "s2_rect_contains"},
-       {"negate", "negate"},
-       {"intdiv", "int_div"},
-       {"lags", "lags"},
-       {"replaceall", "replace_all"},
-       {"farmhash64", "farm_hash64"},
-       {"rightpad", "right_pad"},
-       {"subtractweeks", "subtract_weeks"},
-       {"totimezone", "to_timezone"},
-       {"currentroles", "current_roles"},
-       {"bin", "bin"},
-       {"h3edgelengthkm", "h3_edge_length_km"},
-       {"today", "today"},
-       {"match", "match"},
-       {"torelativemonthnum", "to_relative_month_num"},
-       {"ngramminhashcaseinsensitiveutf8", "ngram_min_hash_case_insensitive_utf8"},
-       {"arraycumsum", "array_cum_sum"},
-       {"arrayfirst", "array_first"},
-       {"emptyarraydatetime", "empty_array_datetime"},
-       {"regiontotopcontinent", "region_to_top_continent"},
-       {"globalnullinignoreset", "global_null_in_ignore_set"},
-       {"emptyarraydate", "empty_array_date"},
-       {"emptyarrayfloat64", "empty_array_float64"},
-       {"h3getunidirectionaledgesfromhexagon", "h3_get_unidirectional_edges_from_hexagon"},
-       {"emptyarrayfloat32", "empty_array_float32"},
-       {"atan", "atan"},
-       {"md4", "md4"},
-       {"queryid", "query_id"},
-       {"emptyarrayuint64", "empty_array_uint64"},
-       {"now", "now"},
-       {"subtractdays", "subtract_days"},
-       {"json_query", "json_query"},
-       {"todatetime32", "to_datetime32"},
-       {"nullif", "null_if"},
-       {"bitand", "bit_and"},
-       {"ngramminhasharg", "ngram_min_hash_arg"},
-       {"boottime", "boot_time"},
-       {"torelativedaynum", "to_relative_day_num"},
-       {"left", "left"},
-       {"asin", "asin"},
-       {"tosecond", "to_second"},
-       {"roundtoexp2", "round_to_exp2"},
-       {"parsedatetime", "parse_datetime"},
-       {"readwktmultipolygon", "read_wkt_multi_polygon"},
-       {"arrayenumerate", "array_enumerate"},
-       {"plus", "plus"},
-       {"dictgetint8ordefault", "dict_get_int8_or_default"},
-       {"reinterpretasint128", "reinterpret_as_int128"},
-       {"arrayconcat", "array_concat"},
-       {"h3pointdistm", "h3_point_dist_m"},
-       {"parsedatetime32besteffort", "parse_datetime32_best_effort"},
-       {"randconstant", "rand_constant"},
-       {"cityhash64", "city_hash64"},
-       {"pointinellipses", "point_in_ellipses"},
-       {"toconcretetype", "to_concrete_type"},
-       {"murmurhash264", "murmur_hash2_64"},
-       {"nullin", "null_in"},
-       {"toint16orzero", "to_int16_or_zero"},
-       {"multifuzzymatchallindices", "multi_fuzzy_match_all_indices"},
-       {"polygonareaspherical", "polygon_area_spherical"},
-       {"dictgetdate", "dict_get_date"},
-       {"dictgetint64", "dict_get_int64"},
-       {"e", "e"},
-       {"bitwrapperfunc", "__bit_wrapper_func"},
-       {"roundage", "round_age"},
-       {"polygonperimeterspherical", "polygon_perimeter_spherical"},
-       {"lgamma", "lgamma"},
-       {"throwif", "throw_if"},
-       {"h3pointdistkm", "h3_point_dist_km"},
-       {"h3isresclassiii", "h3_is_res_class_iii"},
-       {"randstudentt", "rand_student_t"},
-       {"polygonsintersectionspherical", "polygons_intersection_spherical"},
-       {"arraymax", "array_max"},
-       {"serveruuid", "server_uuid"},
-       {"polygonssymdifferencespherical", "polygons_sym_difference_spherical"},
-       {"polygonssymdifferencecartesian", "polygons_sym_difference_cartesian"},
-       {"polygonswithinspherical", "polygons_within_spherical"},
-       {"tostartoffiveminutes", "to_start_of_five_minutes"},
-       {"reinterpretasint32", "reinterpret_as_int32"},
-       {"todecimal64ordefault", "to_decimal64_or_default"},
-       {"arrayresize", "array_resize"},
-       {"joingetornull", "join_get_or_null"},
-       {"encodexmlcomponent", "encode_xml_component"},
-       {"polygonsintersectioncartesian", "polygons_intersection_cartesian"},
-       {"pow", "pow"},
-       {"decrypt", "decrypt"},
-       {"emptyarrayint64", "empty_array_int64"},
-       {"addweeks", "add_weeks"},
-       {"streamingneighbor", "__streaming_neighbor"},
-       {"addtupleofintervals", "add_tuple_of_intervals"},
-       {"toquarter", "to_quarter"},
-       {"yesterday", "yesterday"},
-       {"todateornull", "to_date_or_null"},
-       {"positionutf8", "position_utf8"},
-       {"getscalar", "__get_scalar"},
-       {"ilike", "ilike"},
-       {"touint256ordefault", "to_uint256_or_default"},
-       {"ngramsearchcaseinsensitiveutf8", "ngram_search_case_insensitive_utf8"},
-       {"subtracthours", "subtract_hours"},
-       {"wordshingleminhashcaseinsensitiveutf8", "word_shingle_min_hash_case_insensitive_utf8"},
-       {"pointinpolygon", "point_in_polygon"},
-       {"dictgetuint8ordefault", "dict_get_uint8_or_default"},
-       {"ngramsearchutf8", "ngram_search_utf8"},
-       {"nullinignoreset", "null_in_ignore_set"},
-       {"todayofmonth", "to_day_of_month"},
-       {"mapadd", "map_add"},
-       {"randpoisson", "rand_poisson"},
-       {"randnegativebinomial", "rand_negative_binomial"},
-       {"uuidnumtostring", "uuid_num_to_string"},
-       {"convertcharset", "convert_charset"},
-       {"randbinomial", "rand_binomial"},
-       {"todecimal256ordefault", "to_decimal256_or_default"},
-       {"randbernoulli", "rand_bernoulli"},
-       {"randfisherf", "rand_fisher_f"},
-       {"h3tostring", "h3_to_string"},
-       {"arrayall", "array_all"},
-       {"blockserializedsize", "block_serialized_size"},
-       {"randchisquared", "rand_chi_squared"},
-       {"arrayenumeratedense", "array_enumerate_dense"},
-       {"multisearchallpositionscaseinsensitive", "multi_search_all_positions_case_insensitive"},
-       {"rand64", "rand64"},
-       {"tofloat32ordefault", "to_float32_or_default"},
-       {"radians", "radians"},
-       {"multisearchanycaseinsensitive", "multi_search_any_case_insensitive"},
-       {"emptyarrayint32", "empty_array_int32"},
-       {"randuniform", "rand_uniform"},
-       {"tominute", "to_minute"},
-       {"log2", "log2"},
-       {"mapapply", "map_apply"},
-       {"tuplecast", "tuple_cast"},
-       {"ipv4stringtonumornull", "ipv4_string_to_num_or_null"},
-       {"randomstringutf8", "random_string_utf8"},
-       {"parsedatetime32besteffortorzero", "parse_datetime32_best_effort_or_zero"},
-       {"mapcontainskeylike", "map_contains_key_like"},
-       {"arrayreversefill", "array_reverse_fill"},
-       {"reinterpretasstring", "reinterpret_as_string"},
-       {"reinterpretasuuid", "reinterpret_as_uuid"},
-       {"uuid", "uuid"},
-       {"bitmasktoarray", "bitmask_to_array"},
-       {"tostartofmicrosecond", "to_start_of_microsecond"},
-       {"tostartofhour", "to_start_of_hour"},
-       {"normalizedqueryhash", "normalized_query_hash"},
-       {"reinterpretasdatetime", "reinterpret_as_datetime"},
-       {"tuplehammingdistance", "tuple_hamming_distance"},
-       {"touint8ordefault", "to_uint8_or_default"},
-       {"reinterpretasfloat64", "reinterpret_as_float64"},
-       {"tocolumntypename", "to_column_type_name"},
-       {"reinterpretasfloat32", "reinterpret_as_float32"},
-       {"nowinblock", "now_in_block"},
-       {"sigmoid", "sigmoid"},
-       {"format", "format"},
-       {"emptyarrayint16", "empty_array_int16"},
-       {"replaceregexpone", "replace_regexp_one"},
-       {"reinterpretasint256", "reinterpret_as_int256"},
-       {"fromunixtimestamp64micro", "from_unix_timestamp64_micro"},
-       {"simplejsonextractstring", "simple_json_extract_string"},
-       {"reinterpretasint64", "reinterpret_as_int64"},
-       {"bitmaptoarray", "bitmap_to_array"},
-       {"normalizedqueryhashkeepnames", "normalized_query_hash_keep_names"},
-       {"reinterpretasint16", "reinterpret_as_int16"},
-       {"xor", "xor"},
-       {"addinterval", "add_interval"},
-       {"reinterpretasuint128", "reinterpret_as_uint128"},
-       {"intexp10", "int_exp10"},
-       {"touint16ordefault", "to_uint16_or_default"},
-       {"notin", "not_in"},
-       {"randomstring", "random_string"},
-       {"dicthas", "dict_has"},
-       {"mapsubtract", "map_subtract"},
-       {"linfnormalize", "linf_normalize"},
-       {"hasany", "has_any"},
-       {"tohour", "to_hour"},
-       {"readwktpolygon", "read_wkt_polygon"},
-       {"reinterpretasuint32", "reinterpret_as_uint32"},
-       {"wordshingleminhashargcaseinsensitive", "word_shingle_min_hash_arg_case_insensitive"},
-       {"simplejsonextractfloat", "simple_json_extract_float"},
-       {"filesystemavailable", "filesystem_available"},
-       {"readwktring", "read_wkt_ring"},
-       {"addmilliseconds", "add_milliseconds"},
-       {"cosinedistance", "cosine_distance"},
-       {"minsamplesizecontinous", "min_sample_size_continous"},
-       {"regexpquotemeta", "regexp_quote_meta"},
-       {"emptyarrayuint32", "empty_array_uint32"},
-       {"parsedatetimebesteffort", "parse_datetime_best_effort"},
-       {"uptime", "uptime"},
-       {"arraysort", "array_sort"},
-       {"l1normalize", "l1_normalize"},
-       {"arraydistinct", "array_distinct"},
-       {"addquarters", "add_quarters"},
-       {"ifnotfinite", "if_not_finite"},
-       {"equals", "equals"},
-       {"ifnull", "if_null"},
-       {"regiontoarea", "region_to_area"},
-       {"arrayjaccardindex", "array_jaccard_index"},
-       {"right", "right"},
-       {"dictgetall", "dict_get_all"},
-       {"notequals", "not_equals"},
-       {"s2capunion", "s2_cap_union"},
-       {"factorial", "factorial"},
-       {"tofloat32", "to_float32"},
-       {"arraymin", "array_min"},
-       {"h3getunidirectionaledgeboundary", "h3_get_unidirectional_edge_boundary"},
-       {"multisearchfirstindexcaseinsensitive", "multi_search_first_index_case_insensitive"},
-       {"revision", "revision"},
-       {"visiblewidth", "visible_width"},
-       {"synonyms", "synonyms"},
-       {"randlognormal", "rand_log_normal"},
-       {"multisearchfirstpositionutf8", "multi_search_first_position_utf8"},
-       {"globalin", "global_in"},
-       {"erf", "erf"},
-       {"arraypopback", "array_pop_back"},
-       {"s2rectunion", "s2_rect_union"},
-       {"timezone", "timezone"},
-       {"countequal", "count_equal"},
-       {"isfinite", "is_finite"},
-       {"emptyarrayint8", "empty_array_int8"},
-       {"normalizequerykeepnames", "normalize_query_keep_names"},
-       {"arrayexists", "array_exists"},
-       {"shardnum", "shard_num"},
-       {"hassubstr", "has_substr"},
-       {"streamingnow", "__streaming_now"},
-       {"touint16orzero", "to_uint16_or_zero"},
-       {"multisearchfirstpositioncaseinsensitiveutf8", "multi_search_first_position_case_insensitive_utf8"},
-       {"dictgetipv4", "dict_get_ipv4"},
-       {"replaceregexpall", "replace_regexp_all"},
-       {"like", "like"},
-       {"tostartofmillisecond", "to_start_of_millisecond"},
-       {"fromunixtimestampinjodasyntax", "from_unix_timestamp_in_joda_syntax"},
-       {"tostartoffifteenminutes", "to_start_of_fifteen_minutes"},
-       {"ngramdistanceutf8", "ngram_distance_utf8"},
-       {"min2", "min2"},
-       {"tostartofweek", "to_start_of_week"},
-       {"toint32orzero", "to_int32_or_zero"},
-       {"length", "length"},
-       {"lpnorm", "lp_norm"},
-       {"linfnorm", "linf_norm"},
-       {"datetime64tosnowflake", "datetime64_to_snowflake"},
-       {"touint128ornull", "to_uint128_or_null"},
-       {"tupleminus", "tuple_minus"},
-       {"validatenestedarraysizes", "validate_nested_array_sizes"},
-       {"reinterpretasint8", "reinterpret_as_int8"},
-       {"tolastdayofmonth", "to_last_day_of_month"},
-       {"sign", "sign"},
-       {"upper", "upper"},
-       {"arrayintersect", "array_intersect"},
-       {"notempty", "not_empty"},
-       {"arraysplit", "array_split"},
-       {"dictgetordefault", "dict_get_or_default"},
-       {"tointervalday", "to_interval_day"},
-       {"substringutf8", "substring_utf8"},
-       {"parsedatetimebesteffortornull", "parse_datetime_best_effort_or_null"},
-       {"toint256orzero", "to_int256_or_zero"},
-       {"casewithexpr", "case_with_expr"},
-       {"tanh", "tanh"},
-       {"tostringcuttozero", "to_string_cut_to_zero"},
-       {"arraywithconstant", "array_with_constant"},
-       {"murmurhash232", "murmur_hash2_32"},
-       {"erfc", "erfc"},
-       {"frommodifiedjuliandayornull", "from_modified_julian_day_or_null"},
-       {"replaceone", "replace_one"},
-       {"multisearchallpositionsutf8", "multi_search_all_positions_utf8"},
-       {"asinh", "asinh"},
-       {"detecttonality", "detect_tonality"},
-       {"bitxor", "bit_xor"},
-       {"timezoneof", "timezone_of"},
-       {"greaterorequals", "greater_or_equals"},
-       {"jsonextractuint", "json_extract_uint"},
-       {"h3getfaces", "h3_get_faces"},
-       {"notinignoreset", "not_in_ignore_set"},
-       {"jsonextractarray", "json_extract_array"},
-       {"ngramminhashcaseinsensitive", "ngram_min_hash_case_insensitive"},
-       {"arrayreduce", "array_reduce"},
-       {"todayofyear", "to_day_of_year"},
-       {"l2squarednorm", "l2_squared_norm"},
-       {"h3tocenterchild", "h3_to_center_child"},
-       {"tounixtimestamp", "to_unix_timestamp"},
-       {"toint128ordefault", "to_int128_or_default"},
-       {"degrees", "degrees"},
-       {"tostartofyear", "to_start_of_year"},
-       {"toisoweek", "to_iso_week"},
-       {"regiontopopulation", "region_to_population"},
-       {"geotos2", "geo_to_s2"},
-       {"max2", "max2"},
-       {"upperutf8", "upper_utf8"},
-       {"torelativeminutenum", "to_relative_minute_num"},
-       {"tolowcardinality", "to_low_cardinality"},
-       {"parsedatetime64besteffortornull", "parse_datetime64_best_effort_or_null"},
-       {"tounixtimestamp64milli", "to_unix_timestamp64_milli"},
-       {"hostname", "hostname"},
-       {"tomodifiedjuliandayornull", "to_modified_julian_day_or_null"},
-       {"multiply", "multiply"},
-       {"dictgetuint16ordefault", "dict_get_uint16_or_default"},
-       {"splitbystring", "split_by_string"},
-       {"pi", "pi"},
-       {"version", "version"},
-       {"linfdistance", "linf_distance"},
-       {"torelativehournum", "to_relative_hour_num"},
-       {"datediffwithin", "date_diff_within"},
-       {"datediff", "date_diff"},
-       {"toint64ordefault", "to_int64_or_default"},
-       {"tounixtimestamp64nano", "to_unix_timestamp64_nano"},
-       {"emptyarrayuint8", "empty_array_uint8"},
-       {"multimatchallindices", "multi_match_all_indices"},
-       {"toyyyymmddhhmmss", "to_yyyymmddhhmmss"},
-       {"to_yyyymmddhhmmss", "to_yyyymmddhhmmss"},
-       {"toint128orzero", "to_int128_or_zero"},
-       {"tovalidutf8", "to_valid_utf8"},
-       {"polygonswithincartesian", "polygons_within_cartesian"},
-       {"arrayzip", "array_zip"},
-       {"torelativeyearnum", "to_relative_year_num"},
-       {"ngramsimhash", "ngram_sim_hash"},
-       {"isipaddressinrange", "is_ip_address_in_range"},
-       {"exp2", "exp2"},
-       {"tcpport", "tcp_port"},
-       {"arraylastornull", "array_last_or_null"},
-       {"transform", "transform"},
-       {"bitmapcontains", "bitmap_contains"},
-       {"reinterpretasuint256", "reinterpret_as_uint256"},
-       {"toipv6", "to_ipv6"},
-       {"subtractmicroseconds", "subtract_microseconds"},
-       {"defaultroles", "default_roles"},
-       {"lag", "lag"},
-       {"mapfilter", "map_filter"},
-       {"bitmapcardinality", "bitmap_cardinality"},
-       {"toyyyymmdd", "to_yyyymmdd"},
-       {"to_yyyymmdd", "to_yyyymmdd"},
-       {"lpnormalize", "lp_normalize"},
-       {"l2squareddistance", "l2_squared_distance"},
-       {"readwktpoint", "read_wkt_point"},
-       {"trimboth", "trim_both"},
-       {"s2rectintersection", "s2_rect_intersection"},
-       {"greatest", "greatest"},
-       {"l1distance", "l1_distance"},
-       {"neighbor", "neighbor"},
-       {"arrayreversesort", "array_reverse_sort"},
-       {"trydecrypt", "try_decrypt"},
-       {"tupledividebynumber", "tuple_divide_by_number"},
-       {"touint64ornull", "to_uint64_or_null"},
-       {"defaultvalueoftypename", "default_value_of_type_name"},
-       {"h3distance", "h3_distance"},
-       {"tgamma", "tgamma"},
-       {"s2getneighbors", "s2_get_neighbors"},
-       {"tupledivide", "tuple_divide"},
-       {"arrayfirstindex", "array_first_index"},
-       {"toint32ordefault", "to_int32_or_default"},
-       {"addresstoline", "address_to_line"},
-       {"assumenotnull", "assume_not_null"},
-       {"simplejsonhas", "simple_json_has"},
-       {"arrayjoin", "array_join"},
-       {"bitmapandnotcardinality", "bitmap_andnot_cardinality"},
-       {"detectprogramminglanguage", "detect_programming_language"},
-       {"bitmapxor", "bitmap_xor"},
-       {"gettypeserializationstreams", "get_type_serialization_streams"},
-       {"reinterpretasuint8", "reinterpret_as_uint8"},
-       {"timezoneoffset", "timezone_offset"},
-       {"filesystemcapacity", "filesystem_capacity"},
-       {"arraycast", "array_cast"},
-       {"touint32ordefault", "to_uint32_or_default"},
-       {"startswith", "starts_with"},
-       {"tofloat", "to_float"},
-       {"notilike", "not_ilike"},
-       {"ngramdistance", "ngram_distance"},
-       {"totime", "to_time"},
-       {"multisearchanyutf8", "multi_search_any_utf8"},
-       {"ngramdistancecaseinsensitive", "ngram_distance_case_insensitive"},
-       {"ngramsearchcaseinsensitive", "ngram_search_case_insensitive"},
-       {"arraycumsumnonnegative", "array_cum_sum_non_negative"},
-       {"globalvariable", "global_variable"},
-       {"tointervalweek", "to_interval_week"},
-       {"indexof", "index_of"},
-       {"translate", "translate"},
-       {"h3numhexagons", "h3_num_hexagons"},
-       {"sinh", "sinh"},
-       {"parsedatetimeinjodasyntax", "parse_datetime_in_joda_syntax"},
-       {"tointervalnanosecond", "to_interval_nanosecond"},
-       {"extractall", "extract_all"},
-       {"empty", "empty"},
-       {"simplejsonextractint", "simple_json_extract_int"},
-       {"todatetime", "to_datetime"},
-       {"bar", "bar"},
-       {"h3togeo", "h3_to_geo"},
-       {"indexhint", "index_hint"},
-       {"date_trunc", "date_trunc"},
-       {"date_add", "date_add"},
-       {"date_sub", "date_sub"},
-       {"uniqthetanot", "uniq_theta_not"},
-       {"ceil", "ceil"},
-       {"dictgetuuidordefault", "dict_get_uuid_or_default"},
-       {"utctimestamp", "utc_timestamp"},
-       {"arraylast", "array_last"},
-       {"l2distance", "l2_distance"},
-       {"wordshingleminhash", "word_shingle_min_hash"},
-       {"log10", "log10"},
-       {"mappopulateseries", "map_populate_series"},
-       {"todecimal256ornull", "to_decimal256_or_null"},
-       {"arrayenumerateuniqranked", "array_enumerate_uniq_ranked"},
-       {"makedatetime", "make_datetime"},
-       {"shardcount", "shard_count"},
-       {"h3cellarearads2", "h3_cell_area_rads2"},
-       {"arrayfilter", "array_filter"},
-       {"subtracttupleofintervals", "subtract_tuple_of_intervals"},
-       {"bitnot", "bit_not"},
-       {"jsonextractkeys", "json_extract_keys"},
-       {"addresstolinewithinlines", "address_to_line_with_inlines"},
-       {"l2norm", "l2_norm"},
-       {"uniqthetaintersect", "uniq_theta_intersect"},
-       {"hasall", "has_all"},
-       {"reinterpretasuint64", "reinterpret_as_uint64"},
-       {"tofixedstring", "to_fixed_string"},
-       {"emptyarrayuint16", "empty_array_uint16"},
-       {"tostartofinterval", "to_start_of_interval"},
-       {"tostartofisoyear", "to_start_of_iso_year"},
-       {"h3getindexesfromunidirectionaledge", "h3_get_indexes_from_unidirectional_edge"},
-       {"arraypopfront", "array_pop_front"},
-       {"reinterpretasuint16", "reinterpret_as_uint16"},
-       {"dictgetipv6", "dict_get_ipv6"},
-       {"arrayreversesplit", "array_reverse_split"},
-       {"dictgetdescendants", "dict_get_descendants"},
-       {"crc64", "crc64"},
-       {"least", "least"},
-       {"crc32ieee", "crc32_ieee"},
-       {"todate32ordefault", "to_date32_or_default"},
-       {"arrayavg", "array_avg"},
-       {"isnull", "is_null"},
-       {"tomonday", "to_monday"},
-       {"range", "range"},
-       {"toyyyymm", "to_yyyymm"},
-       {"to_yyyymm", "to_yyyymm"},
-       {"splitbychar", "split_by_char"},
-       {"polygonconvexhullcartesian", "polygon_convex_hull_cartesian"},
-       {"tuplemultiply", "tuple_multiply"},
-       {"tupletonamevaluepairs", "tuple_to_name_value_pairs"},
-       {"touuidordefault", "to_uuid_or_default"},
-       {"greater", "greater"},
-       {"arrayuniq", "array_uniq"},
-       {"less", "less"},
-       {"sha256", "sha256"},
-       {"bittestany", "bit_test_any"},
-       {"arrayreverse", "array_reverse"},
-       {"earliesttimestamp", "earliest_timestamp"},
-       {"tounixtimestamp64micro", "to_unix_timestamp64_micro"},
-       {"arrayfill", "array_fill"},
-       {"arrayproduct", "array_product"},
-       /// case_insensitive_aggregate_functions
-       {"uniqueretract", "__unique_retract"},
-       {"groupuniqarrayretract", "group_uniq_array_retract"},
-       {"minkretract", "__min_k_retract"},
-       {"maxkretract", "__max_k_retract"},
-       {"minretract", "__min_retract"},
-       {"maxretract", "__max_retract"},
-       {"sumretract", "__sum_retract"},
-       {"exponentialtimedecayedcount", "exponential_time_decayed_count"},
-       {"denserank", "dense_rank"},
-       {"rank", "rank"},
-       {"exponentialmovingaverage", "exponential_moving_average"},
-       {"intervallengthsum", "interval_length_sum"},
-       {"singlevalueornull", "single_value_or_null"},
-       {"nothing", "nothing"},
-       {"meanztest", "mean_ztest"},
-       {"rankcorr", "rank_corr"},
-       {"categoricalinformationvalue", "categorical_information_value"},
-       {"sumkahanretract", "__sum_kahan_retract"},
-       {"grouparraymovingavg", "group_array_moving_avg"},
-       {"grouparraymovingsum", "group_array_moving_sum"},
-       {"simplelinearregression", "simple_linear_regression"},
-       {"entropy", "entropy"},
-       {"exponentialtimedecayedsum", "exponential_time_decayed_sum"},
-       {"retention", "retention"},
-       {"histogram", "histogram"},
-       {"uniqueexactretract", "__unique_exact_retract"},
-       {"stochasticlogisticregression", "stochastic_logistic_regression"},
-       {"groupbitmapxor", "group_bitmap_xor"},
-       {"groupbitmapor", "group_bitmap_or"},
-       {"groupbitmap", "group_bitmap"},
-       {"cramersvbiascorrected", "cramers_v_bias_corrected"},
-       {"cramersv", "cramers_v"},
-       {"grouparray", "group_array"},
-       {"p99", "p99"},
-       {"anyheavy", "any_heavy"},
-       {"groupbitand", "group_bit_and"},
-       {"sparkbar", "sparkbar"},
-       {"groupbitor", "group_bit_or"},
-       {"leadinframe", "lead_in_frame"},
-       {"deltasumtimestamp", "delta_sum_timestamp"},
-       {"covarpop", "covar_pop"},
-       {"topkexact", "top_k_exact"},
-       {"p95", "p95"},
-       {"approxtopksum", "approx_top_k_sum"},
-       {"anylast", "any_last"},
-       {"uniquecombined64", "unique_combined64"},
-       {"quantilesexactexclusive", "quantiles_exact_exclusive"},
-       {"uniquetheta", "unique_theta"},
-       {"unique", "unique"},
-       {"laginframe", "lag_in_frame"},
-       {"summapfilteredwithoverflow", "sum_map_filtered_with_overflow"},
-       {"quantilestdigest", "quantiles_t_digest"},
-       {"uniquecombined", "unique_combined"},
-       {"exponentialtimedecayedmax", "exponential_time_decayed_max"},
-       {"minmappedarrays", "min_mapped_arrays"},
-       {"maxk", "max_k"},
-       {"summappedarrays", "sum_mapped_arrays"},
-       {"corr", "corr"},
-       {"mannwhitneyutest", "mann_whitney_utest"},
-       {"stddevpop", "stddev_pop"},
-       {"stddevsamp", "stddev_samp"},
-       {"corrstable", "corr_stable"},
-       {"avgretract", "__avg_retract"},
-       {"covarsampstable", "covar_samp_stable"},
-       {"stddevpopstable", "stddev_pop_stable"},
-       {"maxintersections", "max_intersections"},
-       {"kurtpop", "kurt_pop"},
-       {"stddevsampstable", "stddev_samp_stable"},
-       {"aggthrow", "agg_throw"},
-       {"varsampstable", "var_samp_stable"},
-       {"lastvalue", "last_value"},
-       {"xirr", "xirr"},
-       {"latest", "latest"},
-       {"summapfiltered", "sum_map_filtered"},
-       {"any", "any"},
-       {"groupuniqarray", "group_uniq_array"},
-       {"skewpop", "skew_pop"},
-       {"max", "max"},
-       {"analysisofvariance", "analysis_of_variance"},
-       {"quantilebfloat16weighted", "quantile_b_float16_weighted"},
-       {"countretract", "__count_retract"},
-       {"groupbitxor", "group_bit_xor"},
-       {"quantilesbfloat16", "quantiles_b_float16"},
-       {"boundingratio", "bounding_ratio"},
-       {"skewsamp", "skew_samp"},
-       {"p90", "p90"},
-       {"windowfunnel", "window_funnel"},
-       {"quantilestdigestweighted", "quantiles_t_digest_weighted"},
-       {"quantilebfloat16", "quantile_b_float16"},
-       {"topkexactweighted", "top_k_exact_weighted"},
-       {"quantiletdigestweighted", "quantile_t_digest_weighted"},
-       {"welchttest", "welch_ttest"},
-       {"quantiletdigest", "quantile_t_digest"},
-       {"quantilestimingweighted", "quantiles_timing_weighted"},
-       {"contingency", "contingency"},
-       {"deltasum", "delta_sum"},
-       {"topkweighted", "top_k_weighted"},
-       {"quantiletimingweighted", "quantile_timing_weighted"},
-       {"uniqueexact", "unique_exact"},
-       {"quantiletiming", "quantile_timing"},
-       {"quantilesexactinclusive", "quantiles_exact_inclusive"},
-       {"studentttest", "student_ttest"},
-       {"groupbitmapand", "group_bitmap_and"},
-       {"sequencecount", "sequence_count"},
-       {"avg", "avg"},
-       {"rownumber", "row_number"},
-       {"theilsu", "theils_u"},
-       {"quantilesexacthigh", "quantiles_exact_high"},
-       {"varpopstable", "var_pop_stable"},
-       {"sumwithoverflow", "sum_with_overflow"},
-       {"quantileexactweighted", "quantile_exact_weighted"},
-       {"sequencenextnode", "sequence_next_node"},
-       {"quantileexacthigh", "quantile_exact_high"},
-       {"quantilesexactlow", "quantiles_exact_low"},
-       {"quantilesexactweighted", "quantiles_exact_weighted"},
-       {"quantilestiming", "quantiles_timing"},
-       {"grouparraysample", "group_array_sample"},
-       {"uniquehll12", "unique_hll12"},
-       {"summapwithoverflow", "sum_map_with_overflow"},
-       {"sumcount", "sum_count"},
-       {"earliest", "earliest"},
-       {"quantilesdeterministic", "quantiles_deterministic"},
-       {"sumwithoverflowretract", "__sum_with_overflow_retract"},
-       {"min", "min"},
-       {"varpop", "var_pop"},
-       {"quantileexactlow", "quantile_exact_low"},
-       {"quantiles", "quantiles"},
-       {"sum", "sum"},
-       {"kurtsamp", "kurt_samp"},
-       {"grouparrayinsertat", "group_array_insert_at"},
-       {"quantilesexact", "quantiles_exact"},
-       {"approxtopk", "approx_top_k"},
-       {"exponentialtimedecayedavg", "exponential_time_decayed_avg"},
-       {"firstvalue", "first_value"},
-       {"quantileexact", "quantile_exact"},
-       {"covarpopstable", "covar_pop_stable"},
-       {"sequencematch", "sequence_match"},
-       {"quantilesbfloat16weighted", "quantiles_b_float16_weighted"},
-       {"uniqueupto", "unique_up_to"},
-       {"largesttrianglethreebuckets", "largest_triangle_three_buckets"},
-       {"quantile", "quantile"},
-       {"covarsamp", "covar_samp"},
-       {"varsamp", "var_samp"},
-       {"quantiledeterministic", "quantile_deterministic"},
-       {"quantileexactinclusive", "quantile_exact_inclusive"},
-       {"maxintersectionsposition", "max_intersections_position"},
-       {"maxmappedarrays", "max_mapped_arrays"},
-       {"mink", "min_k"},
-       {"quantileexactexclusive", "quantile_exact_exclusive"},
-       {"sumkahan", "sum_kahan"},
-       {"stochasticlinearregression", "stochastic_linear_regression"},
-       {"count", "count"},
-       {"topk", "top_k"},
-       {"avgweighted", "avg_weighted"}};
-
-String ParserFunction::functionNameFromCamelToSnake(const String & name) const
-{
-    auto it = function_map.find(name);
-    if (it != function_map.end())
-        return it->second;
-
-    String lower_name = Poco::toLower(name);
-    it = case_insensitive_function_map.find(lower_name);
-    if (it != case_insensitive_function_map.end())
-        return it->second;
-
-    return name;
-}
-/// proton: ends
-
 bool ParserFunction::parseImpl(Pos & pos, ASTPtr & node, Expected & expected, [[ maybe_unused ]] bool hint)
 {
     ParserIdentifier id_parser;
@@ -2068,12 +807,11 @@ bool ParserFunction::parseImpl(Pos & pos, ASTPtr & node, Expected & expected, [[
     String function_name_lowercase = Poco::toLower(function_name);
 
     /// proton: starts
-    if (thread_local_is_clickhouse_compatible)
+    if (isClickHouseCompatibleMode())
     {
-        String tmp_name = functionNameFromCamelToSnake(function_name);
-        function_name = tmp_name;
-        auto * tmpnode = dynamic_cast<ASTIdentifier *>(identifier.get());
-        tmpnode->setShortName(function_name);
+        function_name = functionNameFromCamelToSnake(function_name);
+        if (auto * identifier_p = dynamic_cast<ASTIdentifier *>(identifier.get()); identifier_p)
+            identifier_p->setShortName(function_name);
     }
     /// proton: ends
 
@@ -2144,7 +882,7 @@ bool ParserFunction::parseImpl(Pos & pos, ASTPtr & node, Expected & expected, [[
     ParserExpressionList contents(false, is_table_function);
 
     const char * contents_begin = pos->begin;
-    /// proton: starts. support session range comparision expression for table function `session`
+    /// proton: starts. support session range comparison expression for table function `session`
     if (is_table_function && function_name == "session")
     {
         if (!ParserList(
@@ -2192,7 +930,7 @@ bool ParserFunction::parseImpl(Pos & pos, ASTPtr & node, Expected & expected, [[
     /// proton: starts
     /// First determine whether need to be compatible with ClickHouse. Timeplus does not allow the use of syntax such as quantile(0.9)(x).
     /// The parametric aggregate function has two lists (parameters and arguments) in parentheses. Example: quantile(0.9)(x).
-    if (thread_local_is_clickhouse_compatible && allow_function_parameters && pos->type == TokenType::OpeningRoundBracket)
+    if (isClickHouseCompatibleMode() && allow_function_parameters && pos->type == TokenType::OpeningRoundBracket)
     {
         ++pos;
 
@@ -2237,10 +975,6 @@ bool ParserFunction::parseImpl(Pos & pos, ASTPtr & node, Expected & expected, [[
         ++pos;
     }
     /// proton: ends
-
-    /// proton: starts.
-    /// proton: don't support parametric aggregation function any more.
-    /// proton: ends.
 
     auto function_node = std::make_shared<ASTFunction>();
     tryGetIdentifierNameInto(identifier, function_node->name);
