+++ src/Parsers/tests/gtest_compatible_clickhouse_datatype_function.cpp
@@ -6,7 +6,7 @@
 #include <Parsers/ASTLiteral.h>
 
 #include <string_view>
-#include <Common/thread_local_is_clickhouse_compatible.h>
+#include <Common/ClickHouseCompatibleFlag.h>
 
 #include <gtest/gtest.h>
 
@@ -15,7 +15,7 @@ using namespace DB;
 /// clickhouse datatype
 TEST(ParserDataTypeNameTest, TestDatatypeInt8)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Int8);";
     
     ParserFunction parser;
@@ -28,11 +28,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeInt8)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int8");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeUInt8)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS UInt8);";
 
     ParserFunction parser;
@@ -45,11 +46,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeUInt8)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint8");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeInt16)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Int16);";
 
     ParserFunction parser;
@@ -62,11 +64,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeInt16)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int16");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeUInt16)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS UInt16);";
 
     ParserFunction parser;
@@ -79,11 +82,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeUInt16)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint16");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeInt32)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Int32);";
 
     ParserFunction parser;
@@ -96,11 +100,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeInt32)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int32");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeUInt32)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS UInt32);";
 
     ParserFunction parser;
@@ -113,11 +118,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeUInt32)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint32");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeInt64)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Int64);";
 
     ParserFunction parser;
@@ -130,11 +136,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeInt64)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int64");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeUInt64)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS UInt64);";
 
     ParserFunction parser;
@@ -147,11 +154,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeUInt64)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint64");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeInt128)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Int128);";
 
     ParserFunction parser;
@@ -164,11 +172,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeInt128)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int128");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeUInt128)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS UInt128);";
 
     ParserFunction parser;
@@ -181,11 +190,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeUInt128)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint128");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeInt256)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Int256);";
 
     ParserFunction parser;
@@ -198,11 +208,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeInt256)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int256");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeUInt256)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS UInt256);";
 
     ParserFunction parser;
@@ -215,12 +226,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeUInt256)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint256");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeFloat32)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Float32);";
 
     ParserFunction parser;
@@ -233,11 +244,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeFloat32)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "float32");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeFloat64)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Float64);";
 
     ParserFunction parser;
@@ -250,12 +262,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeFloat64)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "float64");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeString)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS String);";
 
     ParserFunction parser;
@@ -268,12 +280,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeString)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "string");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeDate)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Date);";
 
     ParserFunction parser;
@@ -286,11 +298,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDate)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "date");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeDate32)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Date32);";
 
     ParserFunction parser;
@@ -303,11 +316,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDate32)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "date32");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeDateTime)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS DateTime);";
 
     ParserFunction parser;
@@ -320,12 +334,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDateTime)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "datetime");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeBool)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Bool);";
 
     ParserFunction parser;
@@ -338,12 +352,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeBool)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "bool");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeUUID)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS UUID);";
 
     ParserFunction parser;
@@ -356,12 +370,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeUUID)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uuid");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeIPv4)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS IPv4);";
 
     ParserFunction parser;
@@ -374,11 +388,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeIPv4)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "ipv4");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatypeIPv6)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS IPv6);";
 
     ParserFunction parser;
@@ -391,12 +406,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeIPv6)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "ipv6");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeJSON)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS JSON);";
 
     ParserFunction parser;
@@ -409,12 +424,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeJSON)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "json");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeNullable)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Nullable(String));";
 
     ParserFunction parser;
@@ -427,12 +442,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeNullable)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "nullable(string)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeDecimal)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Decimal(10, 3));";
 
     ParserFunction parser;
@@ -445,12 +460,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDecimal)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "decimal(10, 3)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeDecimal32)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Decimal32(3));";
 
     ParserFunction parser;
@@ -463,12 +478,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDecimal32)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "decimal32(3)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeDecimal64)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Decimal64(8));";
 
     ParserFunction parser;
@@ -481,12 +496,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDecimal64)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "decimal64(8)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeDecimal128)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Decimal128(8));";
 
     ParserFunction parser;
@@ -499,12 +514,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDecimal128)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "decimal128(8)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeDecimal256)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Decimal256(8));";
 
     ParserFunction parser;
@@ -517,12 +532,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDecimal256)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "decimal256(8)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeFixedString)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS FixedString(8));";
 
     ParserFunction parser;
@@ -535,12 +550,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeFixedString)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "fixed_string(8)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeDateTime_TimeZone)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS DateTime('Asia/Istanbul'));";
 
     ParserFunction parser;
@@ -553,12 +568,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDateTime_TimeZone)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "datetime('Asia/Istanbul')");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeDateTime64_Precision)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS DateTime64(3));";
 
     ParserFunction parser;
@@ -571,12 +586,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDateTime64_Precision)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "datetime64(3)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeDateTime64_Precision_TimeZone)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS DateTime64(3, 'Asia/Istanbul'));";
 
     ParserFunction parser;
@@ -589,12 +604,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeDateTime64_Precision_TimeZone)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "datetime64(3, 'Asia/Istanbul')");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeEnum)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Enum('One' = 1, 'Two' = 2, 'Three' = 3));";
 
     ParserFunction parser;
@@ -607,12 +622,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeEnum)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "enum('One' = 1, 'Two' = 2, 'Three' = 3)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeArray)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Array(Int32));";
 
     ParserFunction parser;
@@ -625,12 +640,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeArray)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "array(int32)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeMap)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Map(String, Int32));";
 
     ParserFunction parser;
@@ -643,12 +658,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeMap)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "map(string, int32)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeTuple)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Tuple(String, Int32, String));";
 
     ParserFunction parser;
@@ -661,13 +676,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeTuple)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "tuple(string, int32, string)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeLowCardinality)
 {
-
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS LowCardinality(String));";
 
     ParserFunction parser;
@@ -680,13 +694,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeLowCardinality)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "low_cardinality(string)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeNested1)
 {
-
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Nullable(Map(String, Tuple(Array(Nullable(String), Nullable(Int32), Nullable(FixedString(10)), Nullable(DateTime64(3, 'UTC')), LowCardinality(Nullable(Float32)))))));";
 
     ParserFunction parser;
@@ -699,13 +712,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeNested1)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "nullable(map(string, tuple(array(nullable(string), nullable(int32), nullable(fixed_string(10)), nullable(datetime64(3, 'UTC')), low_cardinality(nullable(float32))))))");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeNested2)
 {
-
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Nullable(Array(Tuple(Nullable(Int32), Nullable(String)))));";
 
     ParserFunction parser;
@@ -718,13 +730,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeNested2)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "nullable(array(tuple(nullable(int32), nullable(string))))");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeNested3)
 {
-
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Tuple(Nullable(String), Nullable(Int32), Nullable(DateTime64(3, 'UTC'))));";
 
     ParserFunction parser;
@@ -737,13 +748,12 @@ TEST(ParserDataTypeNameTest, TestDatatypeNested3)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "tuple(nullable(string), nullable(int32), nullable(datetime64(3, 'UTC')))");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatypeNested4)
 {
-
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Map(String, Tuple(Nullable(Int32), Nullable(String))))";
 
     ParserFunction parser;
@@ -756,13 +766,13 @@ TEST(ParserDataTypeNameTest, TestDatatypeNested4)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "map(string, tuple(nullable(int32), nullable(string)))");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 /// timeplus datatype
 TEST(ParserDataTypeNameTest, TestDatatype_int8)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS int8);";
 
     ParserFunction parser;
@@ -775,11 +785,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_int8)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int8");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_uint8)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS uint8);";
 
     ParserFunction parser;
@@ -792,11 +803,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_uint8)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint8");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_int16)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS int16);";
 
     ParserFunction parser;
@@ -809,11 +821,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_int16)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int16");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_uint16)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS uint16);";
 
     ParserFunction parser;
@@ -826,11 +839,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_uint16)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint16");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_int32)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS Int32);";
 
     ParserFunction parser;
@@ -843,11 +857,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_int32)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int32");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_uint32)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS uint32);";
 
     ParserFunction parser;
@@ -860,11 +875,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_uint32)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint32");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_int64)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS int64);";
 
     ParserFunction parser;
@@ -877,11 +893,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_int64)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int64");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_uint64)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS uint64);";
 
     ParserFunction parser;
@@ -894,11 +911,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_uint64)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint64");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_int128)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS int128);";
 
     ParserFunction parser;
@@ -911,11 +929,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_int128)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int128");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_uint128)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS uint128);";
 
     ParserFunction parser;
@@ -928,11 +947,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_uint128)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint128");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_int256)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS int256);";
 
     ParserFunction parser;
@@ -945,11 +965,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_int256)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "int256");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_uint256)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS uint256);";
 
     ParserFunction parser;
@@ -962,12 +983,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_uint256)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uint256");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_float32)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS float32);";
 
     ParserFunction parser;
@@ -980,11 +1001,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_float32)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "float32");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_float64)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS float64);";
 
     ParserFunction parser;
@@ -997,12 +1019,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_float64)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "float64");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_string)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS string);";
 
     ParserFunction parser;
@@ -1015,12 +1037,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_string)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "string");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_date)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS date);";
 
     ParserFunction parser;
@@ -1033,11 +1055,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_date)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "date");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_date32)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS date32);";
 
     ParserFunction parser;
@@ -1050,11 +1073,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_date32)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "date32");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_dateTime)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS datetime);";
 
     ParserFunction parser;
@@ -1067,12 +1091,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_dateTime)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "datetime");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_bool)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS bool);";
 
     ParserFunction parser;
@@ -1085,12 +1109,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_bool)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "bool");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_uuid)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS uuid);";
 
     ParserFunction parser;
@@ -1103,12 +1127,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_uuid)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "uuid");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_ipv4)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS ipv4);";
 
     ParserFunction parser;
@@ -1121,11 +1145,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_ipv4)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "ipv4");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
+
 TEST(ParserDataTypeNameTest, TestDatatype_ipv6)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS ipv6);";
 
     ParserFunction parser;
@@ -1138,12 +1163,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_ipv6)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "ipv6");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_json)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS json);";
 
     ParserFunction parser;
@@ -1156,12 +1181,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_json)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "json");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_nullable)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS nullable(string));";
 
     ParserFunction parser;
@@ -1174,12 +1199,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_nullable)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "nullable(string)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_decimal)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS decimal(10, 3));";
 
     ParserFunction parser;
@@ -1192,12 +1217,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_decimal)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "decimal(10, 3)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_decimal32)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS decimal32(3));";
 
     ParserFunction parser;
@@ -1210,12 +1235,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_decimal32)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "decimal32(3)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_decimal64)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS decimal64(8));";
 
     ParserFunction parser;
@@ -1228,12 +1253,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_decimal64)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "decimal64(8)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_decimal128)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS decimal128(8));";
 
     ParserFunction parser;
@@ -1246,12 +1271,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_decimal128)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "decimal128(8)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_decimal256)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS decimal256(8));";
 
     ParserFunction parser;
@@ -1264,12 +1289,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_decimal256)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "decimal256(8)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_fixed_string)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS fixed_string(8));";
 
     ParserFunction parser;
@@ -1282,12 +1307,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_fixed_string)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "fixed_string(8)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_datetime_timezone)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS datetime('Asia/Istanbul'));";
 
     ParserFunction parser;
@@ -1300,12 +1325,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_datetime_timezone)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "datetime('Asia/Istanbul')");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_datetime64_precision)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS datetime64(3));";
 
     ParserFunction parser;
@@ -1318,12 +1343,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_datetime64_precision)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "datetime64(3)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_datetime64_precision_timezone)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS datetime64(3, 'Asia/Istanbul'));";
 
     ParserFunction parser;
@@ -1336,12 +1361,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_datetime64_precision_timezone)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "datetime64(3, 'Asia/Istanbul')");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_enum)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS enum('One' = 1, 'Two' = 2, 'Three' = 3));";
 
     ParserFunction parser;
@@ -1354,12 +1379,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_enum)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "enum('One' = 1, 'Two' = 2, 'Three' = 3)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_array)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS array(Int32));";
 
     ParserFunction parser;
@@ -1372,12 +1397,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_array)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "array(int32)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_map)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS map(string, int32));";
 
     ParserFunction parser;
@@ -1390,12 +1415,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_map)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "map(string, int32)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_tuple)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS tuple(string, int32, string));";
 
     ParserFunction parser;
@@ -1408,13 +1433,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_tuple)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "tuple(string, int32, string)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_low_cardinality)
 {
-
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS low_cardinality(string));";
 
     ParserFunction parser;
@@ -1427,13 +1451,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_low_cardinality)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "low_cardinality(string)");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_nested1)
 {
-
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS nullable(map(string, tuple(array(nullable(string), nullable(int32), nullable(fixed_string(10)), nullable(datetime64(3, 'UTC')), low_cardinality(nullable(float32)))))));";
 
     ParserFunction parser;
@@ -1446,13 +1469,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_nested1)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "nullable(map(string, tuple(array(nullable(string), nullable(int32), nullable(fixed_string(10)), nullable(datetime64(3, 'UTC')), low_cardinality(nullable(float32))))))");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_nested2)
 {
-
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS nullable(array(tuple(nullable(int32), nullable(string)))));";
 
     ParserFunction parser;
@@ -1465,13 +1487,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_nested2)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "nullable(array(tuple(nullable(int32), nullable(string))))");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_nested3)
 {
-
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS tuple(nullable(string), nullable(int32), nullable(datetime64(3, 'UTC'))));";
 
     ParserFunction parser;
@@ -1484,13 +1505,12 @@ TEST(ParserDataTypeNameTest, TestDatatype_nested3)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "tuple(nullable(string), nullable(int32), nullable(datetime64(3, 'UTC')))");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserDataTypeNameTest, TestDatatype_nested4)
 {
-
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "CAST(id AS map(string, tuple(nullable(int32), nullable(string))))";
 
     ParserFunction parser;
@@ -1503,14 +1523,14 @@ TEST(ParserDataTypeNameTest, TestDatatype_nested4)
     EXPECT_EQ(arg1->name(), "id");
     ASTLiteral * arg2 = function->arguments->children[1]->as<ASTLiteral>();
     EXPECT_EQ(arg2->value, "map(string, tuple(nullable(int32), nullable(string)))");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 
 /// clickhouse function name
 TEST(ParserFunctionNameTest, TestFunction_toUInt16)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "toUInt16('123')";
 
     ParserFunction parser;
@@ -1518,12 +1538,12 @@ TEST(ParserFunctionNameTest, TestFunction_toUInt16)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "to_uint16");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserFunctionNameTest, TestFunction_encodeURLComponent)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "encodeURLComponent('Hello world! How are you?')";
 
     ParserFunction parser;
@@ -1531,12 +1551,12 @@ TEST(ParserFunctionNameTest, TestFunction_encodeURLComponent)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "encode_url_component");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserFunctionNameTest, TestFunction_todAte)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "todAte(now())";
 
     ParserFunction parser;
@@ -1544,12 +1564,12 @@ TEST(ParserFunctionNameTest, TestFunction_todAte)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "to_date");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserFunctionNameTest, TestFunction_generateUUIDv4)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "generateUUIDv4(1)";
 
     ParserFunction parser;
@@ -1557,12 +1577,12 @@ TEST(ParserFunctionNameTest, TestFunction_generateUUIDv4)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "generate_uuidv4");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserFunctionNameTest, TestFunction_stddevSamp)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "stddevSamp(v)";
 
     ParserFunction parser;
@@ -1570,12 +1590,12 @@ TEST(ParserFunctionNameTest, TestFunction_stddevSamp)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "stddev_samp");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserFunctionNameTest, TestFunction_ifNull)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "ifNull(value, 0)";
 
     ParserFunction parser;
@@ -1583,13 +1603,13 @@ TEST(ParserFunctionNameTest, TestFunction_ifNull)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "if_null");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 /// timeplus function name
 TEST(ParserFunctionNameTest, TestFunction_to_uint16)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "to_uint16('123')";
 
     ParserFunction parser;
@@ -1597,12 +1617,12 @@ TEST(ParserFunctionNameTest, TestFunction_to_uint16)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "to_uint16");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserFunctionNameTest, TestFunction_encode_url_component)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "encode_url_component('Hello world! How are you?')";
 
     ParserFunction parser;
@@ -1610,12 +1630,12 @@ TEST(ParserFunctionNameTest, TestFunction_encode_url_component)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "encode_url_component");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserFunctionNameTest, TestFunction_to_date)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "to_date(now())";
 
     ParserFunction parser;
@@ -1623,12 +1643,12 @@ TEST(ParserFunctionNameTest, TestFunction_to_date)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "to_date");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserFunctionNameTest, TestFunction_generate_uuidv4)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "generate_uuidv4(1)";
 
     ParserFunction parser;
@@ -1636,12 +1656,12 @@ TEST(ParserFunctionNameTest, TestFunction_generate_uuidv4)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "generate_uuidv4");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserFunctionNameTest, TestFunction_stddev_samp)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "stddev_samp(v)";
 
     ParserFunction parser;
@@ -1649,12 +1669,12 @@ TEST(ParserFunctionNameTest, TestFunction_stddev_samp)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "stddev_samp");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
 
 TEST(ParserFunctionNameTest, TestFunction_if_null)
 {
-    thread_local_is_clickhouse_compatible = true;
+    setClickHouseCompatibleMode(true);
     String input = "if_null(value, 0)";
 
     ParserFunction parser;
@@ -1662,5 +1682,5 @@ TEST(ParserFunctionNameTest, TestFunction_if_null)
     ASTFunction * function = ast->as<ASTFunction>();
 
     EXPECT_EQ(function->name, "if_null");
-    thread_local_is_clickhouse_compatible = false;
+    setClickHouseCompatibleMode(false);
 }
